Encapsulated all the sync search and the serration search into Resync Class
